--- Assets/Scripts/Player/Camera/Cameras.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ef04d10a244f9f5498291433cd8708e8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
