+++ Assets/Scripts/Player/Camera/Cameras_v2.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ba1543bb1b279e74088bc6b8a21590df
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
