+++ Assets/Scripts/Player/Camera/Cameras_v2_1.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ef04d10a244f9f5498291433cd8708e8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
