+++ Assets/Scripts/Player/General/Player_BlockDetector.cs
@@ -101,7 +101,7 @@ public class Player_BlockDetector : Singleton<Player_BlockDetector>
         }
 
         //Check if something is in the way of movement
-        if (Cameras.Instance.cameraState == CameraState.Forward)
+        if (Cameras_v2.Instance.cameraState == CameraState.Forward)
         {
             //Check if something is in the way
             PerformRaycast_Horizontal(detectorSpot_Horizontal_Front, Vector3.forward);
@@ -115,7 +115,7 @@ public class Player_BlockDetector : Singleton<Player_BlockDetector>
             PerformRaycast_Vertical(detectorSpot_Vertical_Left, Vector3.left);
             PerformRaycast_Vertical(detectorSpot_Vertical_Right, Vector3.right);
         }
-        else if (Cameras.Instance.cameraState == CameraState.Backward)
+        else if (Cameras_v2.Instance.cameraState == CameraState.Backward)
         {
             //Check if something is in the way
             PerformRaycast_Horizontal(detectorSpot_Horizontal_Front, Vector3.back);
@@ -129,7 +129,7 @@ public class Player_BlockDetector : Singleton<Player_BlockDetector>
             PerformRaycast_Vertical(detectorSpot_Vertical_Left, Vector3.right);
             PerformRaycast_Vertical(detectorSpot_Vertical_Right, Vector3.left);
         }
-        else if (Cameras.Instance.cameraState == CameraState.Left)
+        else if (Cameras_v2.Instance.cameraState == CameraState.Left)
         {
             //Check if something is in the way
             PerformRaycast_Horizontal(detectorSpot_Horizontal_Front, Vector3.right);
@@ -143,7 +143,7 @@ public class Player_BlockDetector : Singleton<Player_BlockDetector>
             PerformRaycast_Vertical(detectorSpot_Vertical_Left, Vector3.forward);
             PerformRaycast_Vertical(detectorSpot_Vertical_Right, Vector3.back);
         }
-        else if (Cameras.Instance.cameraState == CameraState.Right)
+        else if (Cameras_v2.Instance.cameraState == CameraState.Right)
         {
             //Check if something is in the way
             PerformRaycast_Horizontal(detectorSpot_Horizontal_Front, Vector3.left);
@@ -327,7 +327,7 @@ public class Player_BlockDetector : Singleton<Player_BlockDetector>
     {
         if (direction == Vector3.forward)
         {
-            switch (Cameras.Instance.cameraState)
+            switch (Cameras_v2.Instance.cameraState)
             {
                 case CameraState.Forward:
                     PlayerManager.Instance.canMove_Forward = false;
@@ -348,7 +348,7 @@ public class Player_BlockDetector : Singleton<Player_BlockDetector>
         }
         if (direction == Vector3.back)
         {
-            switch (Cameras.Instance.cameraState)
+            switch (Cameras_v2.Instance.cameraState)
             {
                 case CameraState.Forward:
                     PlayerManager.Instance.canMove_Back = false;
@@ -369,7 +369,7 @@ public class Player_BlockDetector : Singleton<Player_BlockDetector>
         }
         if (direction == Vector3.left)
         {
-            switch (Cameras.Instance.cameraState)
+            switch (Cameras_v2.Instance.cameraState)
             {
                 case CameraState.Forward:
                     PlayerManager.Instance.canMove_Left = false;
@@ -390,7 +390,7 @@ public class Player_BlockDetector : Singleton<Player_BlockDetector>
         }
         if (direction == Vector3.right)
         {
-            switch (Cameras.Instance.cameraState)
+            switch (Cameras_v2.Instance.cameraState)
             {
                 case CameraState.Forward:
                     PlayerManager.Instance.canMove_Right = false;
@@ -414,7 +414,7 @@ public class Player_BlockDetector : Singleton<Player_BlockDetector>
     {
         if (direction == Vector3.forward)
         {
-            switch (Cameras.Instance.cameraState)
+            switch (Cameras_v2.Instance.cameraState)
             {
                 case CameraState.Forward:
                     PlayerManager.Instance.canMove_Forward = true;
@@ -435,7 +435,7 @@ public class Player_BlockDetector : Singleton<Player_BlockDetector>
         }
         if (direction == Vector3.back)
         {
-            switch (Cameras.Instance.cameraState)
+            switch (Cameras_v2.Instance.cameraState)
             {
                 case CameraState.Forward:
                     PlayerManager.Instance.canMove_Back = true;
@@ -456,7 +456,7 @@ public class Player_BlockDetector : Singleton<Player_BlockDetector>
         }
         if (direction == Vector3.left)
         {
-            switch (Cameras.Instance.cameraState)
+            switch (Cameras_v2.Instance.cameraState)
             {
                 case CameraState.Forward:
                     PlayerManager.Instance.canMove_Left = true;
@@ -477,7 +477,7 @@ public class Player_BlockDetector : Singleton<Player_BlockDetector>
         }
         if (direction == Vector3.right)
         {
-            switch (Cameras.Instance.cameraState)
+            switch (Cameras_v2.Instance.cameraState)
             {
                 case CameraState.Forward:
                     PlayerManager.Instance.canMove_Right = true;
